refactor: move stub class to util package
