--- data-dashboard-backend/src/test/kotlin/org/radarbase/datadashboard/backend/ProjectServiceStub.kt
@@ -1,11 +1,0 @@
-package org.radarbase.datadashboard.backend
-
-import org.radarbase.jersey.service.ProjectService
-
-class ProjectServiceStub : ProjectService {
-    override suspend fun ensureOrganization(organizationId: String) = Unit
-    override suspend fun listProjects(organizationId: String): List<String> = emptyList()
-    override suspend fun projectOrganization(projectId: String): String = "test-organization"
-    override suspend fun ensureProject(projectId: String) = Unit
-    override suspend fun ensureSubject(projectId: String, userId: String) = Unit
-}
