+++ data-dashboard-backend/src/test/kotlin/org/radarbase/datadashboard/backend/util/ProjectServiceStub.kt
@@ -1,0 +1,29 @@
+/*
+ *
+ *  *  Copyright 2026 The Hyve
+ *  *
+ *  *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  *  you may not use this file except in compliance with the License.
+ *  *  You may obtain a copy of the License at
+ *  *
+ *  *    http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *  Unless required by applicable law or agreed to in writing, software
+ *  *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *  See the License for the specific language governing permissions and
+ *  *  limitations under the License.
+ *
+ */
+
+package org.radarbase.datadashboard.backend.util
+
+import org.radarbase.jersey.service.ProjectService
+
+class ProjectServiceStub : ProjectService {
+    override suspend fun ensureOrganization(organizationId: String) = Unit
+    override suspend fun listProjects(organizationId: String): List<String> = emptyList()
+    override suspend fun projectOrganization(projectId: String): String = "test-organization"
+    override suspend fun ensureProject(projectId: String) = Unit
+    override suspend fun ensureSubject(projectId: String, userId: String) = Unit
+}
